Separates main from the app class
