+++ jvm/workbookapp/src/main/kotlin/org/wycliffeassociates/otter/jvm/workbookapp/Main.kt
@@ -1,0 +1,8 @@
+package org.wycliffeassociates.otter.jvm.workbookapp
+
+import org.wycliffeassociates.otter.jvm.workbookapp.ui.OtterApp
+import tornadofx.launch
+
+fun main(args: Array<String>) {
+    launch<OtterApp>(args)
+}
